fix: removed gamecontroller
- removed unused gamecontroller and extracted status rest api to lobbycontroller
- removed gamecontrollertest
